# ENH: improve performance of numpy.core.records.fromarrays