--- a/numpy/core/records.py
+++ b/numpy/core/records.py
@@ -612,7 +612,8 @@ def fromarrays(arrayList, dtype=None, shape=None, formats=None,
         for obj in arrayList:
             if not isinstance(obj, ndarray):
                 raise ValueError("item in the array list must be an ndarray.")
-            formats.append(obj.dtype.str)
+            shape = obj.shape[1:] or ''
+            formats.append('{}{}'.format(shape, obj.dtype.str))
 
     if dtype is not None:
         descr = sb.dtype(dtype)