<!-- Please be sure you are following the instructions in the dev guidelines
http://www.numpy.org/devdocs/dev/gitwash/development_workflow.html
-->

<!-- We'd appreciate it if your commit message is properly formatted
http://www.numpy.org/devdocs/dev/gitwash/development_workflow.html#writing-the-commit-message
-->

### Description
Further improve performance of numpy.core.records.fromarrays: 
* the arrays dtypes were converted to strings, and finally a new comma delimited string was created from those
* the comma separated string was parsed in _format\_parser.\_parseFormats_ basically using regex to extract the formats basically undoing the first step
* the new implementation avoids generating the comma separated string and instead creates the dtype using a list of (field name, field dtype) 
* added benchmark for records

### benchmark

```
       before           after         ratio
     [de28edd8]       [1d1fb908]
     <v1.15.4^0>       <master>  
-     6.79±0.04μs       5.72±0.2μs     0.84  bench_records.Records.time_fromstring_w_dtype
-       177±0.4ms       69.5±0.3ms     0.39  bench_records.Records.time_fromarrays_w_dtype
-         1.29±0s        111±0.6ms     0.09  bench_records.Records.time_fromarrays_formats_as_string
-         1.30±0s       91.1±0.3ms     0.07  bench_records.Records.time_fromarrays_wo_dtype
-      1.28±0.01s       86.0±0.4ms     0.07  bench_records.Records.time_fromarrays_formats_as_list
-      1.12±0.01s       41.4±0.2ms     0.04  bench_records.Records.time_fromstring_formats_as_string
-      1.12±0.01s       16.8±0.2ms     0.01  bench_records.Records.time_fromstring_formats_as_list
```