@danielhrisca 2018-12-20
@eric-wieser 
I've also included a fix for multidimensional arrays passed to fromarrays: use input arrays shape when creating the recarray dtype.