@mattip 2018-12-20
can you point to the benchmarks that improved (or add some)?