@mattip 2018-12-20
> can you point to the benchmarks that improved (or add some)? 

I meant can you add the benchmark to the benchmark suite in the `./benchmark` directory. You should be able to run
```
python runtests.py --bench-compare
```