@danielhrisca 2018-12-20
Some test are failing, however when I run the failed test with numpy 1.13.3 I see the same output. How can this be?