@mattip 2018-12-20
The `test_fromrecords_2d` test was added in 3caf1223b9, in PR #8486 which was part of 1.13. 
On your 1.13.3 that test fails?