@danielhrisca 2018-12-20
@mattip 
sorry for wasting your time. There was an error on my side and the tests are failed with the latest PR code due to the insertion of the array shape in the preparation of the formats.

I think there is an issue with how fromrecords calls fromarrays.

Without considering the input array shapes I get an error with this snippet:

```python
arrays = [
    np.arange(10).reshape((5, 2)),
    np.ones(5) * -1,
]

names = ['f1', 'f2']

arr = np.core.records.fromarrays(arrays, names=names)
```
```
1.13.3
Traceback (most recent call last):
  File "/home/daniel/.config/spyder-py3/temp.py", line 25, in <module>
    arr = np.core.records.fromarrays(arrays, names=names)
  File "/usr/lib/python3/dist-packages/numpy/core/records.py", line 622, in fromarrays
    raise ValueError("array-shape mismatch in array %d" % k)
ValueError: array-shape mismatch in array 1

```
I find this unexpected since the input arrays have the same first dimension

If the shape is considered then the snippet runs ok.