@danielhrisca 2018-12-20
@mattip 
I've added benchmarks for records. Here is the comparison with v1.15.4

```
       before           after         ratio
     [de28edd8]       [5d06372f]
     <v1.15.4^0>       <master>  
-      7.01±0.1μs      5.71±0.02μs     0.81  bench_records.Records.time_fromstring_w_dtype
-        168±10ms       68.5±0.3ms     0.41  bench_records.Records.time_fromarrays_w_dtype
-         1.29±0s        111±0.9ms     0.09  bench_records.Records.time_fromarrays_formats_as_string
-         1.26±0s       93.0±0.8ms     0.07  bench_records.Records.time_fromarrays_wo_dtype
-         1.28±0s       85.2±0.4ms     0.07  bench_records.Records.time_fromarrays_formats_as_list
-         1.11±0s       41.0±0.6ms     0.04  bench_records.Records.time_fromstring_formats_as_string
-         1.11±0s       16.6±0.2ms     0.01  bench_records.Records.time_fromstring_formats_as_list
```