@danielhrisca 2018-12-20
Records seem to be a part of numpy that people don't usually use. I guess nobody ever had the real need to look into performance improvements