@danielhrisca 2018-12-21
I think the original intent in _fromarrays_ was that the input arrays should have the same shape if _dtype_ and _formats_ arguments were _None_. I've added this detail to the docstring

_fromfile_ must be given _dtype_ or _formats_, so I've copied the same check that was performed in _fromstring_