@danielhrisca 2018-12-21
@mattip @eric-wieser 
I think this PR is ready for review