@danielhrisca 2019-01-12
I think it's done now. There were also some trailing white spaces in the benchmark script.